backend/requirements.txt
fastapi==0.109.2
fastapi-utils==0.8.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==2.118.0
PyJWT==2.8.0
pytest==8.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.1.0.post1
aiohttp==3.9.3
typing-extensions==4.9.0
typing-inspect==0.9.0 